aerticket-api/requirements.txt
ansicon==1.89.0
appnope==0.1.4
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
autopep8==2.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
boto3==1.34.160
botocore==1.34.160
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.8.1
decorator==5.1.1
Django==5.0.6
django-baton==3.1.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-filter==24.2
django-js-asset==2.2.0
django-picklefield==3.2
django-q==1.3.9
django-silk==5.1.0
django-storages==1.14.4
django-unixtimestampfield==0.5.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
editor==1.6.6
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.0.1
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
ghp-import==2.1.0
gprof2dot==2024.6.6
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
inquirer==3.2.4
ipykernel==6.29.4
ipython==8.24.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jinxed==1.2.1
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
nest-asyncio==1.6.0
numpy==2.0.0
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.38
pillow==10.3.0
platformdirs==4.2.2
pluggy==0.13.1
prettytable==3.10.0
prompt_toolkit==3.0.45
protobuf==4.23.2
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycountry==24.6.1
pycryptodome==3.22.0
PyGeneratePassword==0.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==26.0.3
razorpay==1.4.2
readchar==4.1.0
redis==3.5.3
requests==2.32.3
rest-framework-simplejwt==0.0.2
runs==1.2.2
s3transfer==0.10.2
scipy==1.14.1
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
stripe==11.1.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
xmltodict==0.14.2
xmod==1.8.1
